On https://www.kaggle.com/bryanpark/sudoku/version/3 there is a file with a lot of sudokus ans its solution. I've created a script that reads the sudoku from the downloaded file, checks that it has a unique solution and then solves it and saves the sudoku on our format (downloaded one, uses 0 for empty clues instead of -). It saves the clues, the number of guesses required to solve the puzzle, if my algorithm solved the sudoku and how many iterations used
Created a generator that starts from an empty board, solves it using random guesses on random cells. Once solved, it starts to remove cells and check if the solution is unique. When the solution stops being unique, it returns the last puzzle with unique solution. It saves it into a file, with some values (how many times a value has to be guessed, if the puzzle has been solved by my algorithm, and how many iterations have been used)

Pending to elaborate a correct way of weighing the cost of solving the puzzle